+++ pkg/block/block_test.go
@@ -89,32 +89,32 @@ func TestUpload(t *testing.T) {
 		{{Name: "a", Value: "3"}},
 		{{Name: "a", Value: "4"}},
 		{{Name: "b", Value: "1"}},
-	}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124)
+	}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124, false)
 	testutil.Ok(t, err)
 	testutil.Ok(t, os.MkdirAll(path.Join(tmpDir, "test", b1.String()), os.ModePerm))
 
 	{
 		// Wrong dir.
-		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "not-existing"))
+		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "not-existing"), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Assert(t, strings.HasSuffix(err.Error(), "/not-existing: no such file or directory"), "")
 	}
 	{
 		// Wrong existing dir (not a block).
-		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test"))
+		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test"), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Equals(t, "not a block dir: ulid: bad data size when unmarshaling", err.Error())
 	}
 	{
 		// Empty block dir.
-		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()))
+		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Assert(t, strings.HasSuffix(err.Error(), "/meta.json: no such file or directory"), "")
 	}
 	e2eutil.Copy(t, path.Join(tmpDir, b1.String(), MetaFilename), path.Join(tmpDir, "test", b1.String(), MetaFilename))
 	{
 		// Missing chunks.
-		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()))
+		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Assert(t, strings.HasSuffix(err.Error(), "/chunks: no such file or directory"), err.Error())
 	}
@@ -122,7 +122,7 @@ func TestUpload(t *testing.T) {
 	e2eutil.Copy(t, path.Join(tmpDir, b1.String(), ChunksDirname, "000001"), path.Join(tmpDir, "test", b1.String(), ChunksDirname, "000001"))
 	{
 		// Missing index file.
-		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()))
+		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Assert(t, strings.HasSuffix(err.Error(), "/index: no such file or directory"), "")
 	}
@@ -130,14 +130,14 @@ func TestUpload(t *testing.T) {
 	testutil.Ok(t, os.Remove(path.Join(tmpDir, "test", b1.String(), MetaFilename)))
 	{
 		// Missing meta.json file.
-		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()))
+		err := Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Assert(t, strings.HasSuffix(err.Error(), "/meta.json: no such file or directory"), "")
 	}
 	e2eutil.Copy(t, path.Join(tmpDir, b1.String(), MetaFilename), path.Join(tmpDir, "test", b1.String(), MetaFilename))
 	{
 		// Full block.
-		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String())))
+		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()), metadata.NoneFunc))
 		testutil.Equals(t, 4, len(bkt.Objects()))
 		testutil.Equals(t, 3751, len(bkt.Objects()[path.Join(b1.String(), ChunksDirname, "000001")]))
 		testutil.Equals(t, 401, len(bkt.Objects()[path.Join(b1.String(), IndexFilename)]))
@@ -187,7 +187,7 @@ func TestUpload(t *testing.T) {
 	}
 	{
 		// Test Upload is idempotent.
-		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String())))
+		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, "test", b1.String()), metadata.NoneFunc))
 		testutil.Equals(t, 4, len(bkt.Objects()))
 		testutil.Equals(t, 3751, len(bkt.Objects()[path.Join(b1.String(), ChunksDirname, "000001")]))
 		testutil.Equals(t, 401, len(bkt.Objects()[path.Join(b1.String(), IndexFilename)]))
@@ -201,9 +201,9 @@ func TestUpload(t *testing.T) {
 			{{Name: "a", Value: "3"}},
 			{{Name: "a", Value: "4"}},
 			{{Name: "b", Value: "1"}},
-		}, 100, 0, 1000, nil, 124)
+		}, 100, 0, 1000, nil, 124, false)
 		testutil.Ok(t, err)
-		err = Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b2.String()))
+		err = Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b2.String()), metadata.NoneFunc)
 		testutil.NotOk(t, err)
 		testutil.Equals(t, "empty external labels are not allowed for Thanos block.", err.Error())
 		testutil.Equals(t, 4, len(bkt.Objects()))
@@ -226,9 +226,9 @@ func TestDelete(t *testing.T) {
 			{{Name: "a", Value: "3"}},
 			{{Name: "a", Value: "4"}},
 			{{Name: "b", Value: "1"}},
-		}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124)
+		}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124, false)
 		testutil.Ok(t, err)
-		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b1.String())))
+		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b1.String()), metadata.NoneFunc))
 		testutil.Equals(t, 4, len(bkt.Objects()))
 
 		// Full delete.
@@ -243,9 +243,9 @@ func TestDelete(t *testing.T) {
 			{{Name: "a", Value: "3"}},
 			{{Name: "a", Value: "4"}},
 			{{Name: "b", Value: "1"}},
-		}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124)
+		}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124, false)
 		testutil.Ok(t, err)
-		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b2.String())))
+		testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b2.String()), metadata.NoneFunc))
 		testutil.Equals(t, 5, len(bkt.Objects()))
 
 		// Remove meta.json and check if delete can delete it.
@@ -297,12 +297,12 @@ func TestMarkForDeletion(t *testing.T) {
 				{{Name: "a", Value: "3"}},
 				{{Name: "a", Value: "4"}},
 				{{Name: "b", Value: "1"}},
-			}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124)
+			}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124, false)
 			testutil.Ok(t, err)
 
 			tcase.preUpload(t, id, bkt)
 
-			testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, id.String())))
+			testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, id.String()), metadata.NoneFunc))
 
 			c := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
 			err = MarkForDeletion(ctx, log.NewNopLogger(), bkt, id, "", c)
@@ -353,12 +353,12 @@ func TestMarkForNoCompact(t *testing.T) {
 				{{Name: "a", Value: "3"}},
 				{{Name: "a", Value: "4"}},
 				{{Name: "b", Value: "1"}},
-			}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124)
+			}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 124, false)
 			testutil.Ok(t, err)
 
 			tcase.preUpload(t, id, bkt)
 
-			testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, id.String())))
+			testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, id.String()), metadata.NoneFunc))
 
 			c := promauto.With(nil).NewCounter(prometheus.CounterOpts{})
 			err = MarkForNoCompact(ctx, log.NewNopLogger(), bkt, id, metadata.ManualNoCompactReason, "", c)
@@ -366,4 +366,42 @@ func TestMarkForNoCompact(t *testing.T) {
 			testutil.Equals(t, float64(tcase.blocksMarked), promtest.ToFloat64(c))
 		})
 	}
+}
+
+// TestHashDownload uploads an empty block to in-memory storage
+// and tries to download it to the same dir. It should not try
+// to download twice.
+func TestHashDownload(t *testing.T) {
+	defer testutil.TolerantVerifyLeak(t)
+
+	ctx := context.Background()
+
+	tmpDir, err := ioutil.TempDir("", "test-block-download")
+	testutil.Ok(t, err)
+	defer func() { testutil.Ok(t, os.RemoveAll(tmpDir)) }()
+
+	bkt := objstore.NewInMemBucket()
+	b1, err := e2eutil.CreateBlockWithTombstone(ctx, tmpDir, []labels.Labels{
+		{{Name: "a", Value: "1"}},
+	}, 100, 0, 1000, labels.Labels{{Name: "ext1", Value: "val1"}}, 42, true)
+	testutil.Ok(t, err)
+	testutil.Ok(t, Upload(ctx, log.NewNopLogger(), bkt, path.Join(tmpDir, b1.String()), metadata.SHA256Func))
+	testutil.Equals(t, 4, len(bkt.Objects()))
+
+	m, err := metadata.ReadFromDir(path.Join(tmpDir, b1.String()))
+	testutil.Ok(t, err)
+	for _, fl := range m.Thanos.Files {
+		testutil.Assert(t, fl.Hash != nil, "expected a hash for %s but got nil", fl.RelPath)
+	}
+	err = Download(ctx, log.NewNopLogger(), bkt, m.ULID, path.Join(tmpDir, b1.String()))
+	testutil.Ok(t, err)
+	testutil.Equals(t, 0, bkt.GetDownloadedCount(m.ULID.String()+"/"+"index"))
+	testutil.Equals(t, 1, bkt.GetDownloadedCount(m.ULID.String()+"/"+"meta.json"))
+	testutil.Equals(t, 0, bkt.GetDownloadedCount(m.ULID.String()+"/"+"chunks/000001"))
+
+	err = Download(ctx, log.NewNopLogger(), bkt, m.ULID, path.Join(tmpDir, b1.String()))
+	testutil.Ok(t, err)
+	testutil.Equals(t, 0, bkt.GetDownloadedCount(m.ULID.String()+"/"+"index"))
+	testutil.Equals(t, 2, bkt.GetDownloadedCount(m.ULID.String()+"/"+"meta.json"))
+	testutil.Equals(t, 0, bkt.GetDownloadedCount(m.ULID.String()+"/"+"chunks/000001"))
 }